+++ test/cmdTest.js
@@ -1,23 +1,23 @@
-var assert = require( "assert" );
-var path = require( "path" );
-var childProcess = require( "child_process" );
-var fs = require( "fs" );
+import assert from "assert";
+import path from "path";
+import childProcess from "child_process";
+import fs from "fs";
+import pkg from "../package.json" with { type: "json" };
 
 describe( "CLI (version, help)", function() {
 	it( "outputs version", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --version`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
-		let pkg = require("../package.json");
 		assert.equal( output.toString().trim(), pkg.version );
 	});
 
 	it( "outputs help with --help", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --help`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.ok( output.toString().indexOf("usage: glyphhanger") > -1 );
@@ -26,7 +26,7 @@ describe( "CLI (version, help)", function() {
 	it( "outputs help (no arguments)", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.ok( output.toString().indexOf("usage: glyphhanger") > -1 );
@@ -38,7 +38,7 @@ describe( "CLI (urls)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/multiple/one.html`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+61-63" );
@@ -47,7 +47,7 @@ describe( "CLI (urls)", function() {
 	it( "can use multiple urls", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js test/multiple/one.html test/multiple/two.html`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+61-66" );
@@ -56,7 +56,7 @@ describe( "CLI (urls)", function() {
 	it( "can use spider", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js test/urls/root.html --spider`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+41-47" );
@@ -65,7 +65,7 @@ describe( "CLI (urls)", function() {
 	it( "can use spider with limit", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js test/urls/root.html --spider-limit=2`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+41-44" );
@@ -75,7 +75,7 @@ describe( "CLI (urls)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/test-emoji.html`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.deepEqual( output.toString().trim(), "U+31-34,U+41-44,U+1F4A9");
@@ -85,7 +85,7 @@ describe( "CLI (urls)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/test-emoji.txt`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.deepEqual( output.toString().trim(), "U+31-34,U+41-44,U+1F4A9");
@@ -95,7 +95,7 @@ describe( "CLI (urls)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/json/families.html --json`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		let json = JSON.parse(output.toString().trim());
@@ -111,7 +111,7 @@ describe( "CLI (urls)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/json/families.html --family`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		let json = JSON.parse(output.toString().trim());
@@ -127,7 +127,7 @@ describe( "CLI (urls)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/json/families.html --family='A Web Font'`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.deepEqual( output.toString().trim(), "U+67-69" );
@@ -137,7 +137,7 @@ describe( "CLI (urls)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/json/families.html --family='A Web Font, monospace'`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.deepEqual( output.toString().trim(), "U+64-69" );
@@ -148,7 +148,7 @@ describe( "CLI (whitelist)", function() {
 	it( "--whitelist without other arguments (outputs range)", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --whitelist=ABCD`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+41-44" );
@@ -157,7 +157,7 @@ describe( "CLI (whitelist)", function() {
 	it( "--whitelist --string (outputs string)", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --whitelist=ABCD --string`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "ABCD" );
@@ -166,7 +166,7 @@ describe( "CLI (whitelist)", function() {
 	it( "--US_ASCII (outputs code points)", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --US_ASCII`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+20-7E" );
@@ -175,7 +175,7 @@ describe( "CLI (whitelist)", function() {
 	it( "--whitelist --US_ASCII (outputs code points)", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --whitelist=ABCD --US_ASCII`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+20-7E" );
@@ -184,7 +184,7 @@ describe( "CLI (whitelist)", function() {
 	it( "--whitelist='Unicode Range' --US_ASCII (outputs code points)", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --whitelist="U+09" --US_ASCII`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+9,U+20-7E" );
@@ -195,7 +195,7 @@ describe( "CLI (css)", function() {
 	it( "--css --whitelist='Unicode Range'", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --whitelist="U+09" --css`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.ok( output.toString().indexOf(`@font-face {
@@ -206,7 +206,7 @@ describe( "CLI (css)", function() {
 	it( "--css --whitelist='String'", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`node cmd.js --whitelist="abcd" --css`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.ok( output.toString().indexOf(`@font-face {
@@ -218,7 +218,7 @@ describe( "CLI (css)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/multiple/one.html --css`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.ok( output.toString().indexOf(`@font-face {
@@ -230,7 +230,7 @@ describe( "CLI (css)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/json/families.html --family='A Web Font, monospace' --css`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.ok( output.toString().indexOf(`@font-face {
@@ -243,7 +243,7 @@ describe( "CLI (css)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/test-onlyVisible.html --onlyVisible`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+31-34" );
@@ -253,7 +253,7 @@ describe( "CLI (css)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/test-cssSelector.html --cssSelector="pre"`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+20,U+61,U+63,U+65,U+67-69,U+6D-70,U+73,U+74" );
@@ -263,7 +263,7 @@ describe( "CLI (css)", function() {
 		this.timeout( 10000 );
 
 		let output = childProcess.execSync(`node cmd.js test/test-cssSelectorVisible.html --cssSelector="pre" --onlyVisible`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "U+20,U+61,U+63,U+65,U+67-69,U+6D-70,U+73,U+74" );
@@ -272,7 +272,7 @@ describe( "CLI (css)", function() {
 	it( "System input and --jsdom", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`echo "<b>ih</b>" | node cmd.js --jsdom --string`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "hi" );
@@ -281,7 +281,7 @@ describe( "CLI (css)", function() {
 	it( "System input and --jsdom (not HTML, gets transformed to HTML for JSDOM)", function () {
 		this.timeout( 10000 );
 		let output = childProcess.execSync(`echo "ih" | node cmd.js --jsdom --string`, {
-			cwd: path.resolve(__dirname, "..")
+			cwd: path.resolve(import.meta.dirname, "..")
 		});
 
 		assert.equal( output.toString().trim(), "hi" );
